Fixes bug in ThenToPromise
Same as the bug in CatchRetToPromise.

If Then return a void, the ThenToPromise
cannot correctly take void as a parameter.

The solution is to use partial specialization
when the Catch returns void, return CPromiseFail
directly.

It also auto detect if the PromiseType need bool
and If ParamType is a Promise or Params.